Add some specs that fail to exhibit the problem
I thought one of these would fail, but it didn't.  The problem I saw in
my other tests was that an object with the same object_id had a
different __mock_proxy.object_id, which implied that Mock::Space was
different for each of the threads.

Have to think through how to cause this to be the case...Just wanted to
document what I know so far.